Temporary fix for strobe synchronization
The PulseStrobe::kPrimingPulseFPS constant should be being picked up from the config environment in the camera 2 processs, but it not currently.  So, this is a temporary fix to ensure that the constants is the same for both processes.  The difference (15 versus 30 FPS) was causing the strobe pulses to be un-synchronized with the camera shutter.